requirements/base.txt
django-classy-tags>=0.8.0
